IndexedDB/submissions/Opera/invalid/idbfactory_deleteDatabase.htm

<!DOCTYPE html>
<title>IDBFactory.deleteDatabase() - request has no source </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support.js"></script>

<p>Bogus test</p>
<script type="text/javascript">
    // XXX: THIS TEST IS BOGUS. deleteDatabase returns void, not a request...
    var db,
      t = async_test();

    var open_rq = createdb(t);
    open_rq.onsuccess = function(e) {
        db = e.target.result;

        var req = window.indexedDB.deleteDatabase(db.name);
        req.onerror = fail(t, "Error event from deleteDatabase");
        req.onsuccess = t.step_func(function(e) {
            assert_equals(e.target.source, null);
            t.done();
        });
    }

</script>

<div id="log"></div>
